pyproject.toml
[project]
name = "cuteninja"
version = "0.1.0"
description = "KDL as HTML replacement for Jinja2 templates"
readme = "README.md"
requires-python = ">=3.13"
license-files = ["LICENSE"]
classifiers = [
  "Development Status :: 3 - Alpha",
  "Intended Audience :: Developers",
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Text Processing :: Markup :: HTML",
]
dependencies = ["kdl-rs-py>=0.1.0"]

[project.optional-dependencies]
dev = ["pytest>=7.0", "pytest-cov>=4.0"]

[build-system]
requires = ["setuptools>=65.0", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
packages = ["cuteninja"]
